Б19: модуль form.js - из него вынес функционал загрузки фотографии - в  photo.js
